requirements.txt
google-generativeai
python-dotenv
pandas
gspread
oauth2client
flask
requests
